Updated Crowe.Exercise.ConsoleApp project.
Updated appsettings.json with a comment that if 'WriteToConsole' is true, this means that HelloWorld message will be written to database. Added 'GetMessageEndpoint' to call WebApi (Get) Message method. Updated ConsoleApp Program.cs to include the GetMessage method which handles calling and getting the message data from the WebApi (Get) Message method. Added printing of the returned response message to the console via the Main entry point method.